implement relays for data and logs
