+++ brewblox_history/__main__.py
@@ -2,8 +2,9 @@
 Example of how to import and use the brewblox service
 """
 
-from brewblox_history import builder, influx, logcollector, sse
 from brewblox_service import brewblox_logger, events, scheduler, service
+
+from brewblox_history import builder, influx, relays, sse
 
 LOGGER = brewblox_logger(__name__)
 
@@ -13,6 +14,9 @@ def create_parser(default_name='history'):
     parser.add_argument('--broadcast-exchange',
                         help='Eventbus exchange to which device services broadcast their state. [%(default)s]',
                         default='brewcast')
+    parser.add_argument('--logging-exchange',
+                        help='Eventbus exchange to which device services broadcast their logs. [%(default)s]',
+                        default='logcast')
     parser.add_argument('--skip-influx-config',
                         help='Skip configuring Influx database after each connection. '
                         'This is useful when using custom configuration.',
@@ -23,16 +27,19 @@ def create_parser(default_name='history'):
 def main():
     app = service.create_app(parser=create_parser())
 
-    # Setup history functionality
     scheduler.setup(app)
     events.setup(app)
     influx.setup(app)
     builder.setup(app)
     sse.setup(app)
-    logcollector.setup(app)
+    relays.setup(app)
 
-    influx.get_relay(app).subscribe(
+    relays.get_data_relay(app).subscribe(
         exchange_name=app['config']['broadcast_exchange'],
+        routing='#'
+    )
+    relays.get_log_relay(app).subscribe(
+        exchange_name=app['config']['logging_exchange'],
         routing='#'
     )
