+++ brewblox_history/influx.py
@@ -1,5 +1,4 @@
 import asyncio
-import collections
 import datetime
 from concurrent.futures import CancelledError
 from typing import Iterator
@@ -7,39 +6,26 @@ from typing import Iterator
 from aiohttp import web
 from aiohttp.client_exceptions import ClientConnectionError
 from aioinflux import InfluxDBClient
-from brewblox_service import brewblox_logger, events, features, scheduler
+from brewblox_service import brewblox_logger, features, scheduler
 
 LOGGER = brewblox_logger(__name__)
-routes = web.RouteTableDef()
 
 
 INFLUX_HOST = 'influx'
-FLAT_SEPARATOR = '/'
 WRITE_INTERVAL_S = 1
 RECONNECT_INTERVAL_S = 1
 MAX_PENDING_POINTS = 5000
 
 DEFAULT_DATABASE = 'brewblox'
-DEFAULT_RETENTION = '1w'
+LOG_DATABASE = 'brewblox_logs'
 
+DEFAULT_RETENTION = '1w'
 DOWNSAMPLE_INTERVALS = ['10s', '1m', '10m', '1h']
 DOWNSAMPLE_RETENTION = ['INF', 'INF', 'INF', 'INF']
 
 
 def setup(app):
-    features.add(app, InfluxWriter(app))
-    features.add(app, EventRelay(app))
     features.add(app, QueryClient(app))
-
-    app.router.add_routes(routes)
-
-
-def get_writer(app) -> 'InfluxWriter':
-    return features.get(app, InfluxWriter)
-
-
-def get_relay(app) -> 'EventRelay':
-    return features.get(app, EventRelay)
 
 
 def get_client(app) -> 'QueryClient':
@@ -52,20 +38,21 @@ class QueryClient(features.ServiceFeature):
     No attempt is made to buffer or reschedule queries if the database is unreachable.
     """
 
-    def __init__(self, app: web.Application=None):
+    def __init__(self, app: web.Application):
         super().__init__(app)
         self._client: InfluxDBClient = None
 
     async def startup(self, app: web.Application):
         await self.shutdown()
-        self._client = InfluxDBClient(host=INFLUX_HOST, loop=app.loop)
+        self._client = InfluxDBClient(host=INFLUX_HOST, loop=app.loop, db=DEFAULT_DATABASE)
 
     async def shutdown(self, *_):
         if self._client:
             await self._client.close()
             self._client = None
 
-    async def query(self, query: str, database: str=DEFAULT_DATABASE, **kwargs):
+    async def query(self, query: str, database: str=None, **kwargs):
+        database = database or DEFAULT_DATABASE
         return await self._client.query(query, db=database, **kwargs)
 
 
@@ -77,18 +64,20 @@ class InfluxWriter(features.ServiceFeature):
 
     If the database is unreachable when write_soon() is called,
     the data points are kept until the database is available again.
+
+    Offers optional downsampling for all measurements in the database.
     """
 
     def __init__(self,
-                 app: web.Application=None,
-                 database: str=DEFAULT_DATABASE,
-                 retention: str=DEFAULT_RETENTION,
+                 app: web.Application,
+                 database: str=None,
+                 retention: str=None,
                  downsampling: bool=True):
         super().__init__(app)
 
         self._pending = []
-        self._database = database
-        self._retention = retention
+        self._database = database or DEFAULT_DATABASE
+        self._retention = retention or DEFAULT_RETENTION
         self._downsampling = downsampling
         self._task: asyncio.Task = None
         self._skip_config = False
@@ -210,143 +199,3 @@ class InfluxWriter(features.ServiceFeature):
         if len(self._pending) >= MAX_PENDING_POINTS:
             LOGGER.warn(f'Downsampling pending points...')
             self._pending = self._pending[::2]
-
-
-class EventRelay(features.ServiceFeature):
-    """Writes all data from specified event queues to the database.
-
-    After a subscription is set, it will relay all incoming messages.
-
-    When relaying, the data dict is flattened.
-    The first part of the routing key is considered the controller name,
-    and becomes the InfluxDB measurement name.
-
-    All subsequent routing key components are considered to be sub-set indicators of the controller.
-    If the routing key is controller1.block1.sensor1, we consider this as being equal to:
-
-        'controller1': {
-            'block1': {
-                'sensor1': <event data>
-            }
-        }
-
-    Data in sub-dicts (including those implied by routing key) is flattened.
-    The key name will be the path to the sub-dict, separated by /.
-
-    If we'd received an event where:
-
-        routing_key = 'controller1.block1.sensor1'
-        data = {
-            settings: {
-                'setting': 'setting'
-            },
-            values: {
-                'value': 'val',
-                'other': 1
-            }
-        }
-
-    it would be flattened to:
-
-        {
-            'block1/sensor1/settings/setting': 'setting',
-            'block1/sensor1/values/value': 'val',
-            'block1/sensor1/values/other': 1
-        }
-
-    If the event data is not a dict, but a string, it is first converted to:
-
-        {
-            'text': <string data>
-        }
-
-    This dict is then flattened.
-    """
-
-    def __init__(self, app: web.Application):
-        super().__init__(app, startup=features.Startup.MANUAL)
-        self._listener = events.get_listener(app)
-        self._writer = get_writer(app)
-
-    async def startup(self, *_):
-        pass
-
-    async def shutdown(self, *_):
-        pass
-
-    def subscribe(self, *args, **kwargs):
-        """Adds relay behavior to subscription.
-
-        All arguments to this function are passed to brewblox_service.events.subscribe()
-        """
-        kwargs['on_message'] = self._on_event_message
-        self._listener.subscribe(*args, **kwargs)
-
-    def _flatten(self, d, parent_key='', sep='/'):
-        items = []
-        for k, v in d.items():
-            new_key = f'{parent_key}{sep}{k}' if parent_key else str(k)
-
-            if isinstance(v, list):
-                v = {li: lv for li, lv in enumerate(v)}
-
-            if isinstance(v, collections.MutableMapping):
-                items.extend(self._flatten(v, new_key, sep=sep).items())
-            else:
-                items.append((new_key, v))
-        return dict(items)
-
-    async def _on_event_message(self,
-                                subscription: events.EventSubscription,
-                                routing: str,
-                                message: dict):
-        # Routing is formatted as controller name followed by active sub-index
-        # A complete push of the controller state is routed as just the controller name
-        routing_list = routing.split('.')
-
-        # Convert textual messages to a dict before flattening
-        if isinstance(message, str):
-            message = dict(text=message)
-
-        parent = FLAT_SEPARATOR.join(routing_list[1:])
-        data = self._flatten(message, parent_key=parent, sep=FLAT_SEPARATOR)
-
-        await self._writer.write_soon(measurement=routing_list[0], fields=data)
-
-
-@routes.post('/subscribe')
-async def add_subscription(request: web.Request) -> web.Response:
-    """
-    ---
-    tags:
-    - History
-    summary: Add a new event subscription
-    description: All messages matching the subscribed topic will be relayed to the database.
-    operationId: history.subscribe
-    produces:
-    - application/json
-    parameters:
-    -
-        in: body
-        name: body
-        description: subscription
-        required: true
-        schema:
-            type: object
-            properties:
-                exchange:
-                    type: string
-                    example: brewblox
-                routing:
-                    type: string
-                    example: controller.#
-    """
-    args = await request.json()
-    exchange = args['exchange']
-    routing = args['routing']
-
-    get_relay(request.app).subscribe(
-        exchange_name=exchange,
-        routing=routing)
-
-    return web.Response()
